+++ packages/react-components/src/components/legacy-grid.tsx
@@ -1,6 +1,6 @@
 import { gifPaginator, GifsResult } from '@giphy/js-fetch-api'
 import { IGif, IUser } from '@giphy/js-types'
-import { getGifHeight } from '@giphy/js-util'
+import Bricks from 'bricks.js'
 import { css, cx } from 'emotion'
 import React, { GetDerivedStateFromProps, PureComponent, ReactType } from 'react'
 import { debounce } from 'throttle-debounce'
@@ -8,7 +8,10 @@ import Observer from '../util/observer'
 import FetchError from './fetch-error'
 import Gif, { EventProps, GifOverlayProps } from './gif'
 import Loader from './loader'
-import MasonryGrid from './masonry-grid'
+
+const loaderHiddenCss = css`
+    opacity: 0;
+`
 
 type Props = {
     className?: string
@@ -66,12 +69,49 @@ class Grid extends PureComponent<Props, State> {
         return null
     }
 
+    setBricks() {
+        const { columns, gutter } = this.props
+        // bricks
+        this.bricks = Bricks({
+            container: this.el!,
+            packed: `data-packed-${columns}`,
+            sizes: [{ columns, gutter }],
+        })
+    }
+
     componentDidMount() {
+        this.setBricks()
         this.onFetch()
     }
 
     componentWillUnmount() {
         this.unmounted = true
+    }
+
+    componentDidUpdate(prevProps: Props, prevState: State) {
+        const { gifs } = this.state
+        const { gifWidth } = this.state
+
+        const numberOfOldGifs = prevState.gifs.length
+        const numberOfNewGifs = gifs.length
+
+        if (prevState.gifWidth !== gifWidth && numberOfOldGifs > 0) {
+            const { columns } = this.props
+            if (columns !== prevProps.columns) {
+                this.setBricks()
+            }
+            this.bricks.pack()
+        }
+
+        if (prevState.gifs !== gifs) {
+            if (numberOfNewGifs > numberOfOldGifs && numberOfOldGifs > 0) {
+                // we just added new gifs
+                this.bricks.update()
+            } else {
+                // we changed existing gifs or removed a gif
+                this.bricks.pack()
+            }
+        }
     }
 
     onLoaderVisible = (isVisible: boolean) => {
@@ -119,16 +159,13 @@ class Grid extends PureComponent<Props, State> {
             overlay,
             hideAttribution,
             noResultsMessage,
-            columns,
-            gutter,
         } = this.props
         const { gifWidth, gifs, isError, isDoneFetching } = this.state
         const showLoader = fetchGifs && !isDoneFetching
         const isFirstLoad = gifs.length === 0
-        const itemHeights = gifs.map(gif => getGifHeight(gif, gifWidth))
         return (
             <div className={className}>
-                <MasonryGrid itemHeights={itemHeights} itemWidth={gifWidth} columns={columns} gutter={gutter}>
+                <div ref={c => (this.el = c)}>
                     {gifs.map(gif => (
                         <Gif
                             gif={gif}
@@ -143,8 +180,8 @@ class Grid extends PureComponent<Props, State> {
                             hideAttribution={hideAttribution}
                         />
                     ))}
-                </MasonryGrid>
-                {!showLoader && gifs.length === 0 && noResultsMessage}
+                    {!showLoader && gifs.length === 0 && noResultsMessage}
+                </div>
                 {isError ? (
                     <FetchError onClick={this.onFetch} />
                 ) : (
@@ -158,9 +195,5 @@ class Grid extends PureComponent<Props, State> {
         )
     }
 }
-
-const loaderHiddenCss = css`
-    opacity: 0;
-`
 
 export default Grid
