+++ packages/react-components/src/index.tsx
@@ -3,10 +3,10 @@ import 'intersection-observer'
 
 export { default as Attribution } from './components/attribution'
 export { default as AttributionOverlay } from './components/attribution/overlay'
-export { default as GridBricks } from './components/bricks-grid'
 export { default as Carousel } from './components/carousel'
 export { default as Gif, GifOverlayProps, PingbackContext } from './components/gif'
 export { default as Grid } from './components/grid'
+export { default as LegacyGrid } from './components/legacy-grid'
 
 const { version } = require('../package.json')
